+++ libraries/glassfish_helper.rb
@@ -1,0 +1,35 @@
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+class RealityForge #nodoc
+  module GlassFish #nodoc
+    class << self
+      # Return the current domain name
+      #
+      # The domain is typically set when configuration run starts
+      #
+      def get_current_domain(node)
+        domain_key = node.run_state['glassfish_domain']
+        raise 'No domain specified' unless domain_key
+        domain_key
+      end
+
+      # Set the current domain name
+      #
+      def set_current_domain(node, domain_key)
+        node.run_state['glassfish_domain'] = domain_key
+      end
+    end
+  end
+end